desyra/back_end_desyra/requirements.txt
aniso8601==7.0.0
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dateparser==1.2.2
defusedxml==0.7.1
Django==5.2.3
django-cors-headers==4.7.0
django-jet-reboot==1.3.10
djangorestframework==3.16.0
dnspython==2.7.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==3.2.3
idna==3.10
jet-bridge-base==1.12.0
jet-django==1.12.0
mysqlclient==2.2.7
oauthlib==3.3.1
pillow==11.2.1
promise==2.3
prompt-toolkit==2.0.9
psutil==7.0.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.1.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
Rx==1.6.3
shop_python==0.0.5
six==1.17.0
social-auth-core==4.7.0
SQLAlchemy==1.4.54
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wcwidth==0.2.13
